+++ packages/headless/src/features/facets/automatic-facet-set/automatic-facet-set-slice.ts
@@ -1,5 +1,6 @@
 import {createReducer} from '@reduxjs/toolkit';
 import {FacetValue} from '../../../product-listing.index';
+import {deselectAllBreadcrumbs} from '../../breadcrumb/breadcrumb-actions';
 import {restoreSearchParameters} from '../../search-parameters/search-parameter-actions';
 import {executeSearch} from '../../search/search-actions';
 import {
@@ -64,6 +65,11 @@ export const automaticFacetSetReducer = createReducer(
 
           state.set[field] = {response};
         }
+      })
+      .addCase(deselectAllBreadcrumbs, (state) => {
+        Object.values(state.set).forEach(({response}) => {
+          response.values.forEach((value) => (value.state = 'idle'));
+        });
       });
   }
 );